Backend and Frontend Changes

+++ Backend/AuthService/authservice/src/main/java/com/authservice/authservice/Repository/AuthRepository.java
@@ -4,7 +4,7 @@ import com.shared.sharedlib.Entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import java.math.BigDecimal;
 
-public interface AuthRepository extends JpaRepository<User, BigDecimal> {
+public interface AuthRepository extends JpaRepository<User, Long> {
 
 
     User findByUserName(String userName);
